fix: headway widget
